+++ src/cpp/spell/mana_potion.cc
@@ -12,12 +12,15 @@ ManaPotion::ManaPotion(Player& player) : Spell(player) {
 void ManaPotion::Cast() {
   Spell::Cast();
   const double kCurrentPlayerMana = player.stats.mana;
+
   // todo check for the randomize values option
-  const int kManaGain = player.rng.range(min_mana_gain, max_mana_gain);
+  const double kManaGain = player.rng.range(min_mana_gain, max_mana_gain);
   player.stats.mana = std::min(player.stats.max_mana, kCurrentPlayerMana + kManaGain);
+
   if (player.recording_combat_log_breakdown) {
     player.AddIterationDamageAndMana(name, kManaGain, 0);
   }
+
   if (player.ShouldWriteToCombatLog()) {
     player.CombatLog("Player gains " + DoubleToString(round(player.stats.mana - kCurrentPlayerMana)) + " mana from " +
                      name + " (" + DoubleToString(round(kCurrentPlayerMana)) + " -> " +
